refactor(analysis): isolate adaptive line indexing
